+++ sympy/matrices/expressions/fourier.py
