+++ sympy/matrices/expressions/tests/test_fourier.py
@@ -1,5 +1,5 @@
 from sympy import S, I, ask, Q, Abs, simplify, exp, sqrt
-from sympy.matrices.expressions.fft import DFT, IDFT
+from sympy.matrices.expressions.fourier import DFT, IDFT
 from sympy.matrices import det, Matrix, Identity
 from sympy.abc import n, i, j
 def test_dft():
